user view load data
